# src/docs/swagger.yaml
definitions:
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
  description: API de controle financeiro pessoal.
  title: SaldoZen API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credenciais
        in: body
        name: credentials
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: erro
          schema:
            type: string
        "401":
          description: erro
          schema:
            type: string
      summary: Login
      tags:
      - auth
  /refresh:
    post:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Revalidar token
      tags:
      - Auth
...
  /users/{id}:
    get:
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Detalhes do usuário
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"